backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asttokens==3.0.0
attrs==24.2.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
debugpy==1.8.12
decorator==5.2.1
distro==1.9.0
dj-rest-auth==7.0.1
Django==4.2.20
django-allauth==65.8.1
django-cors-headers==4.7.0
django-extensions==3.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
exceptiongroup==1.2.2
executing==2.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.4
importlib_metadata==6.7.0  # 8.x는 3.10 이상 필요
ipykernel==6.29.5
ipython==8.12.0            # 3.10 미만에서 호환되는 마지막
jedi==0.19.2
jiter==0.8.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
openai==0.28
outcome==1.3.0.post0
packaging==24.2
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==6.1.1
pure_eval==0.2.3
pycparser==2.22
pydantic==1.10.13           # 2.x는 3.10 이상 필요
Pygments==2.19.1
PyJWT==2.9.0
PySocks==1.7.1; sys_platform == 'win32'  # Windows 전용 패키지에 조건 추가
python-dateutil==2.9.0.post0

python-dotenv==0.21.1
numpy==2.0.2

pywin32==308; sys_platform == "win32"
pyzmq==26.2.0
requests==2.32.3
selenium==4.30.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
Wikipedia-API==0.7.3
wsproto==1.2.0
zipp==3.15.0                 # 최신 버전은 Python >=3.10 요구
psycopg2-binary
gunicorn